* [apiclient] reorganize models and managers
