--- apiclient/harvester_api/__init__.pyi
@@ -1,3 +1,0 @@
-from .api import HarvesterAPI
-
-API = HarvesterAPI
