+++ apiclient/harvester_api/managers/__init__.py
@@ -1,0 +1,34 @@
+from .hosts import HostManager
+from .images import ImageManager
+from .volumes import VolumeManager
+from .backups import BackupManager, VirtualMachineSnapshotManager
+from .keypairs import KeypairManager
+from .settings import SettingManager
+from .networks import NetworkManager
+from .templates import TemplateManager
+from .supportbundles import SupportBundleManager
+from .storageclasses import StorageClassManager
+from .clusternetworks import ClusterNetworkManager
+from .volumesnapshots import VolumeSnapshotManager
+from .virtualmachines import VirtualMachineManager
+# Not available in dashboard
+from .internals import VersionManager, UpgradeManager
+from .longhorns import LonghornReplicaManager, LonghornVolumeManager, LonghornBackupVolumeManager
+
+__all__ = [
+ "HostManager",
+ "ImageManager",
+ "VolumeManager",
+ "KeypairManager",
+ "SettingManager",
+ "NetworkManager",
+ "TemplateManager",
+ "SupportBundleManager",
+ "VolumeSnapshotManager",
+ "ClusterNetworkManager",
+ "VirtualMachineManager",
+ "StorageClassManager",
+ "BackupManager", "VirtualMachineSnapshotManager",
+ "VersionManager", "UpgradeManager",
+ "LonghornReplicaManager", "LonghornVolumeManager", "LonghornBackupVolumeManager",
+]
