+++ apiclient/harvester_api/managers/managers.pyi
