+++ apiclient/harvester_api/models/backups.py
@@ -1,0 +1,58 @@
+from copy import deepcopy
+
+
+class RestoreSpec:
+    def __init__(self, new_vm, vm_name=None, namespace=None, delete_volumes=None):
+        self.new_vm = new_vm
+        self.vm_name = vm_name
+        self.namespace = namespace
+        self.delete_volumes = delete_volumes
+
+    def __repr__(self):
+        return (f"{__class__.__name__}({self.new_vm},"
+                f" {self.vm_name}, {self.namespace}, {self.delete_volumes})")
+
+    def to_dict(self, name, namespace, existing_vm):
+        data = {
+            "type": "harvesterhci.io.virtualmachinerestore",
+            "metadata": {
+                "generateName": f"restore-{name}-",
+                "name": "",
+                "namespace": self.namespace or namespace
+            },
+            "spec": {
+                "target": {
+                    "apiGroup": "kubevirt.io",
+                    "kind": "VirtualMachine",
+                    "name": self.vm_name
+                },
+                "virtualMachineBackupName": name,
+                "virtualMachineBackupNamespace": namespace
+            }
+        }
+        spec = data['spec']
+        if self.new_vm:
+            spec['newVM'] = self.new_vm
+        else:
+            spec['deletionPolicy'] = "delete" if self.delete_volumes else "retain"
+            spec['target']['name'] = existing_vm
+
+        return deepcopy(data)
+
+    @classmethod
+    def for_new(cls, vm_name, namespace=None):
+        return cls(True, vm_name, namespace)
+
+    @classmethod
+    def for_existing(cls, delete_volumes=True):
+        return cls(False, delete_volumes=delete_volumes)
+
+
+class SnapshotRestoreSpec(RestoreSpec):
+    @classmethod
+    def for_new(cls, vm_name):
+        return super().for_new(vm_name, None)
+
+    @classmethod
+    def for_existing(cls):
+        return super().for_existing(False)
