+++ apiclient/harvester_api/models/base.py
@@ -1,0 +1,4 @@
+from typing import Final
+
+MGMT_NETID: Final = object()
+DEFAULT_STORAGE_CLS: Final = "harvester-longhorn"
