+++ apiclient/harvester_api/models/virtualmachines.py
@@ -1,67 +1,9 @@
 from copy import deepcopy
-from json import dumps, loads
 
 import yaml
 
-MGMT_NETID = object()
-DEFAULT_STORAGE_CLS = "harvester-longhorn"
-
-
-class RestoreSpec:
-    def __init__(self, new_vm, vm_name=None, namespace=None, delete_volumes=None):
-        self.new_vm = new_vm
-        self.vm_name = vm_name
-        self.namespace = namespace
-        self.delete_volumes = delete_volumes
-
-    def __repr__(self):
-        return (f"{__class__.__name__}({self.new_vm},"
-                f" {self.vm_name}, {self.namespace}, {self.delete_volumes})")
-
-    def to_dict(self, name, namespace, existing_vm):
-        data = {
-            "type": "harvesterhci.io.virtualmachinerestore",
-            "metadata": {
-                "generateName": f"restore-{name}-",
-                "name": "",
-                "namespace": self.namespace or namespace
-            },
-            "spec": {
-                "target": {
-                    "apiGroup": "kubevirt.io",
-                    "kind": "VirtualMachine",
-                    "name": self.vm_name
-                },
-                "virtualMachineBackupName": name,
-                "virtualMachineBackupNamespace": namespace
-            }
-        }
-        spec = data['spec']
-        if self.new_vm:
-            spec['newVM'] = self.new_vm
-        else:
-            spec['deletionPolicy'] = "delete" if self.delete_volumes else "retain"
-            spec['target']['name'] = existing_vm
-
-        return deepcopy(data)
-
-    @classmethod
-    def for_new(cls, vm_name, namespace=None):
-        return cls(True, vm_name, namespace)
-
-    @classmethod
-    def for_existing(cls, delete_volumes=True):
-        return cls(False, delete_volumes=delete_volumes)
-
-
-class SnapshotRestoreSpec(RestoreSpec):
-    @classmethod
-    def for_new(cls, vm_name):
-        return super().for_new(vm_name, None)
-
-    @classmethod
-    def for_existing(cls):
-        return super().for_existing(False)
+from .base import MGMT_NETID, DEFAULT_STORAGE_CLS
+from .volumes import VolumeSpec
 
 
 class VMSpec:
@@ -431,193 +373,3 @@ class VMSpec:
         obj._cloudinit_vol = dict(disk=devices['disks'][-1], volume=volumes[-1])
         obj._data = data
         return obj
-
-
-class VolumeSpec:
-    volume_mode = "Block"
-    _data = None
-
-    def __init__(self, size, storage_cls=None, description=None, annotations=None):
-        self.access_modes = ["ReadWriteMany"]
-
-        self.size = size
-        self.storage_cls = storage_cls
-        self.description = description
-        self.annotations = annotations
-
-    def to_dict(self, name, namespace, image_id=None):
-        annotations = self.annotations or dict()
-        size = f"{self.size}Gi" if isinstance(self.size, (int, float)) else self.size
-        data = {
-            "type": "persistentvolumeclaim",
-            "metadata": {
-                "namespace": namespace,
-                "name": name,
-                "annotations": annotations
-            },
-            "spec": {
-                "accessModes": self.access_modes,
-                "resources": {
-                    "requests": dict(storage=size)
-                },
-                "volumeMode": self.volume_mode
-            }
-        }
-
-        if image_id is not None:
-            annotations.update({"harvesterhci.io/imageId": image_id})
-        if self.description is not None:
-            annotations.update({"field.cattle.io/description": self.description})
-        if self.storage_cls is not None:
-            data['spec']['storageClassName'] = self.storage_cls
-
-        if self._data:
-            self._data['metadata'].update(data['metadata'])
-            self._data['spec'].update(data['spec'])
-            return deepcopy(self._data)
-
-        return deepcopy(data)
-
-    @classmethod
-    def from_dict(cls, data):
-        spec, metadata = data.get('spec', {}), data.get('metadata', {})
-        annotations = metadata.get('annotations')
-        size = spec['resources']['requests']['storage']
-        storage_cls = spec.get('storageClassName')
-
-        obj = cls(size, storage_cls, annotations.get("field.cattle.io/description"))
-        obj.annotations = annotations
-        obj._data = data
-
-        return obj
-
-
-class BaseSettingSpec:
-    """ Base class for instance check and create"""
-    _name = ""  # to point out the name of setting
-    _default = False  # to point out to use default value
-
-    def __init__(self, value=None):
-        self.value = value or dict()
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}({self.value})"
-
-    @property
-    def use_default(self):
-        return self._default
-
-    @use_default.setter
-    def use_default(self, val):
-        self._default = val
-
-    def to_dict(self, data):
-        return dict(value=dumps(self.value))
-
-    @classmethod
-    def from_dict(cls, data):
-        for c in cls.__subclasses__():
-            if c._name == data.get('metadata', {}).get('name'):
-                return c.from_dict(data)
-        return cls(data)
-
-
-class BackupTargetSpec(BaseSettingSpec):
-    _name = "backup-target"
-
-    @property
-    def type(self):
-        return self.value.get('type')
-
-    def clear(self):
-        self.value = dict()
-
-    @classmethod
-    def from_dict(cls, data):
-        value = loads(data.get('value', "{}"))
-        return cls(value)
-
-    @classmethod
-    def S3(cls, bucket, region, access_id, access_secret, endpoint="", virtual_hosted=None):
-        data = {
-            "type": "s3",
-            "endpoint": endpoint,
-            "bucketName": bucket,
-            "bucketRegion": region,
-            "accessKeyId": access_id,
-            "secretAccessKey": access_secret
-        }
-        if virtual_hosted is not None:
-            data['virtualHostedStyle'] = virtual_hosted
-        return cls(data)
-
-    @classmethod
-    def NFS(cls, endpoint):
-        return cls(dict(type="nfs", endpoint=endpoint))
-
-
-class StorageNetworkSpec(BaseSettingSpec):
-    _name = "storage-network"
-
-    @classmethod
-    def disable(cls):
-        obj = cls()
-        obj.use_default = True
-        return obj
-
-    @classmethod
-    def enable_with(cls, vlan_id, cluster_network, ip_range, *excludes):
-        return cls({
-            "vlan": vlan_id,
-            "clusterNetwork": cluster_network,
-            "range": ip_range,
-            "exclude": excludes
-        })
-
-    @classmethod
-    def from_dict(cls, data):
-        value = loads(data.get('value', "{}"))
-        return cls(value)
-
-    def to_dict(self, data):
-        if self.use_default or not self.value:
-            return dict(value=None)
-        return super().to_dict(data)
-
-
-class OverCommitConfigSpec(BaseSettingSpec):
-    _name = 'overcommit-config'
-
-    @property
-    def cpu(self):
-        return self.value['cpu']
-
-    @cpu.setter
-    def cpu(self, val):
-        self.value['cpu'] = val
-
-    @property
-    def memory(self):
-        return self.value['memory']
-
-    @memory.setter
-    def memory(self, val):
-        self.value['memory'] = val
-
-    @property
-    def storage(self):
-        return self.value['storage']
-
-    @storage.setter
-    def storage(self, val):
-        self.value['storage'] = val
-
-    @classmethod
-    def from_dict(cls, data):
-        value = loads(data.get('value', '{}'))
-        return cls(value)
-
-    def to_dict(self, data):
-        if self.use_default or not self.value:
-            return dict(value=data['default'])
-        return super().to_dict(data)
